+++ src/shared/event/server_event.rs
@@ -31,6 +31,7 @@ use crate::shared::{
         replicon_server::RepliconServer,
     },
     postcard_utils,
+    protocol::ProtocolHasher,
     replication::client_ticks::ClientTicks,
     replicon_tick::RepliconTick,
 };
@@ -173,6 +174,10 @@ impl ServerEventAppExt for App {
         deserialize: EventDeserializeFn<ClientReceiveCtx, E>,
     ) -> &mut Self {
         debug!("registering event `{}`", any::type_name::<E>());
+
+        self.world_mut()
+            .resource_mut::<ProtocolHasher>()
+            .add_server_event::<E>();
 
         let event_fns = EventFns::new(serialize, deserialize);
         let event = ServerEvent::new(self, channel, event_fns);
@@ -812,7 +817,7 @@ impl BufferedServerEvents {
                                 event.send(server, client_entity, ticks)?;
                             } else {
                                 debug!(
-                                    "ignoring broadcast for channel {} for non-replicated client `{client_entity}`",
+                                    "ignoring broadcast for channel {} for non-authorized client `{client_entity}`",
                                     event.channel_id
                                 );
                             }
@@ -829,7 +834,7 @@ impl BufferedServerEvents {
                                 event.send(server, client_entity, ticks)?;
                             } else {
                                 debug!(
-                                    "ignoring broadcast except `{entity}` for channel {} for non-replicated client `{client_entity}`",
+                                    "ignoring broadcast except `{entity}` for channel {} for non-authorized client `{client_entity}`",
                                     event.channel_id
                                 );
                             }
@@ -842,7 +847,7 @@ impl BufferedServerEvents {
                                     event.send(server, client_entity, ticks)?;
                                 } else {
                                     error!(
-                                        "ignoring direct event for non-replicated client `{client_entity}`, \
+                                        "ignoring direct event for non-authorized client `{client_entity}`, \
                                          mark it as independent to allow this"
                                     );
                                 }