+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Authorization system. By default we just verify compatibility between client and server, but it's customizable via `RepliconSharedPlugin::auth_method`.
 - Configurable `SendRate` for deterministic replication. Use `SendRate::Once` to send only the initial value, or `SendRate::Periodic` to only sync the state periodically.
 - `AppRuleExt::replicate_with_priority` to configure replication rule priority.
 - `DisconnectRequest` event to queue a disconnection for a specific client on the server.
@@ -18,6 +19,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `AppRuleExt::replicate_with` now accepts `IntoReplicationRule` trait that allows to define rules with multiple components.
 - Rename `GroupReplication` into `BundleReplication`.
+- Rename `ReplicatedClient` into `AuthorizedClient`.
 - Rename `AppRuleExt::replicate_group` into `AppRuleExt::replicate_bundle`.
 - Rename `replication_registry::despawn_recursive` into `replication_registry::despawn`.
 - Rename `shared::event::trigger` module into `shared::event::remote_targets`.
@@ -37,6 +39,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - `WriteCtx::commands`. You can now insert and remove components directly through `DeferredEntity`.
 - `RepliconClient::receive` and `RepliconServer::receive` are now private since they should only be called internally by Replicon.
+- `ServerPlugin::replicate_after_connect`. Use `RepliconSharedPlugin::auth_method` with `AuthMethod::Custom` instead.
 - Deprecated methods.
 
 ## [0.33.0] - 2025-04-27